Barbican service setup needed in agent configuration role
Issues:
Fixes #5

Problem:
For installing and configuring the agent, we often need to setup
barbican as the certificate manager. This functionality needs to be
added in the agent configuration role.

Analysis:
Modified the configure_lbaasv2_agent role task to add the barbican
service to the agent's ini file when configuring the agent.

Tests:
Tested on TLC deployment in boulder lab.