Prepare ansible scripts for newton installation of driver and agent
Issues:
Fixes #10

Problem:
There is one major change that needs to be done to be able to install
the agent and driver for OpenStack Newton. It appears that setting the
service_provider in the neutron_lbaas.conf file is deprecated. The
ansible playbook will then put the service_provider configuration in the
neutron.conf file.

Analysis:
Changed the configure driver playbook to set the service provider in the
neutron.conf file instead of neutron_lbaas.conf

Tests:
Testing with TLC manually