Boolean conditional checks for 'when' should not use jinja2 format
Issues:
Fixes #11

Problem:
As of ansible release 2.3.0, a warning is printed when using the jinja2
format for checking conditionals in a when statement. The suggestion is
to coerce a variable to a boolean and do the check in this way: 'when:
my_var|bool' or 'when: not my_var|bool'.

Analysis:
Modified all when statements to make this coercion and check in the way
outlined in the problem statement above.

Tests:
Tested on TLC deployment with newton branch. These same changes are
going into liberty and mitaka. It will then merge up to newton.