Make agent installation perform an optional upgrade when using pip
Issues:
Fixes #16

Problem:
When the agent is installed, we perform an upgrade with pip. This has a
side-effect of upgrading dependencies to the latest versions, even if it
is strictly not necessary for testing. We should be able to optionally
upgrade with pip, but the default behavior will be to not upgrade.

Analysis:
Added a default pip_upgrade variable and set it to 'False'. Set a new
condition for when use_pip and the pip_upgrade are set to 'True'.

Tests:
Ran deployment on existing liberty stack.