Make git a dependency of each role
Issues:
Fixes #30

Problem:
Git must be installed as a dependency of each role where we use pip to install
f5 code. This is because pip often installs from the github repository. We can
add a set of preparatory steps, such as installing git, where we can collect
such tasks as dependencies.

Analysis:
Added a new role, f5_ansible_prepare_host, which functions as a dependency for
each role while requires it.

Tests:
Testing in the Jenkins build.