Create playbook to install ml2 mechanism driver
Issues:
Fixes #33

Problem:
We need to be able to install the f5 ml2 mechanism driver on a node that
contains neutron. This can be done with a single role, to pip install
the library, configure neutron, and restart the neutron server.

Analysis:
Created the playbook and the vars file.

Tests:
Testing with ansible container.