In ml2 playbook vars file, update ml2 branch to be stable/newton instead of stable/mitaka
Issues:
Fixes #37

Problem:
In the stable/newton branch of this repo, the stable/mitaka branch is
referenced for the location of the ml2 driver, but it should be changed
to stable/newton, since the two branches will drift quickly in that
repo. These are just default values in variables, but we should modify
it for clarity.

Analysis:
Modified the default variable to point to stable/newton instead of
stable/mitaka.

Tests: